+++ app/app_components/dashboardModule/services/notificationService.js
@@ -1,14 +1,47 @@
 'use strict';
 angular.module('dashboardModule')
-    .factory('notificationService', ['$http', function ($http) {
+    .factory('notificationService', ['$http', '$interval', '$q', function ($http, $interval, $q) {
         var url = '/api/notifications';
+        var notifications = {};
+        var intervalPromise = null;
 
-        function getNotifications() {
-            return $http.get(url).then(
+        function startPolling() {
+            fetchNotifications();
+
+            if (intervalPromise !== null) {
+                $interval.cancel(intervalPromise);
+                intervalPromise = null;
+            }
+            intervalPromise = $interval(fetchNotifications, 10e3);
+        }
+
+        function stopPolling() {
+            if (intervalPromise !== null) {
+                $interval.cancel(intervalPromise);
+                intervalPromise = null;
+            }
+            notifications = {};
+        }
+
+        function fetchNotifications() {
+            $http.get(url).then(
                 function (response) {
-                    return response.data.items;
+                    notifications = response.data;
+                },
+                function (error) {
+                    if (error === null) {
+                        notifications = {};
+                    } else if (error.status === 403) {
+                        stopPolling();
+                    }
                 }
             );
+        }
+
+        function getNotifications() {
+            var deferred = $q.defer();
+            deferred.resolve(notifications);
+            return deferred.promise;
         }
 
         function markNotificationRead(notificationId) {
@@ -19,9 +52,16 @@ angular.module('dashboardModule')
             return $http.delete(url);
         }
 
+        // start polling after page reload. cookie MIGHT be available
+        startPolling();
+
         return {
             getNotifications: getNotifications,
             markNotificationRead: markNotificationRead,
-            markAllNotificationsRead: markAllNotificationsRead
+            markAllNotificationsRead: markAllNotificationsRead,
+
+            startPolling: startPolling,
+            stopPolling: stopPolling
         };
-    }]);
+    }])
+;